fix(curators): Curator.timeout_s is float not int
Timeouts are routinely sub-second or fractional across the two providers
we'll implement (Ollama warm-path is <1s, Anthropic SDK lets us pass
fractional connection/read timeouts). Declaring int would force each
impl to truncate or cast. Caught in Task 1 code review before Tasks 2
and 3 hardcoded the wrong type downstream.

+++ src/claude_almanac/curators/base.py
@@ -12,8 +12,8 @@ from typing import Protocol, runtime_checkable
 
 @runtime_checkable
 class Curator(Protocol):
-    name: str       # provider id: "ollama" | "anthropic_sdk"
-    model: str      # "gemma3:4b", "claude-haiku-4-5-20251001", etc.
-    timeout_s: int
+    name: str         # provider id: "ollama" | "anthropic_sdk"
+    model: str        # "gemma3:4b", "claude-haiku-4-5-20251001", etc.
+    timeout_s: float  # seconds; fractional values are permitted
 
     def invoke(self, system_prompt: str, user_turn: str) -> str: ...
